Links configured, coordinates type fix, BinMetadata fields nullability fix

+++ app/src/main/java/ru/alexdeadman/cfttesttask/data/binlist/retrofit/binmetadata/Country.kt
@@ -8,6 +8,6 @@ data class Country(
     @SerializedName("name") var name: String? = null,
     @SerializedName("emoji") var emoji: String? = null,
     @SerializedName("currency") var currency: String? = null,
-    @SerializedName("latitude") var latitude: Int? = null,
-    @SerializedName("longitude") var longitude: Int? = null
+    @SerializedName("latitude") var latitude: Double? = null,
+    @SerializedName("longitude") var longitude: Double? = null
 )
